requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-core
python_dotenv
pandas
langchain-experimental
langgraph
IPython
pygraphviz
streamlit
psycopg2
psycopg2-binary